+++ demisto_sdk/commands/update_release_notes/update_rn.py
@@ -115,6 +115,8 @@ class UpdateRN:
             # incident fields and indicator fields are using the same scheme.
             elif 'IncidentFields' in file_path:
                 _file_type = 'IncidentFields'
+            elif 'IndicatorTypes' in file_path:
+                _file_type = 'IndicatorTypes'
             elif 'IncidentTypes' in file_path:
                 _file_type = 'IncidentTypes'
             elif 'Classifiers' in file_path:
@@ -194,6 +196,7 @@ class UpdateRN:
         classifier_header = False
         layout_header = False
         inc_types_header = False
+        ind_types_header = False
         for k, v in changed_items.items():
             if k == 'N/A':
                 continue
@@ -201,37 +204,42 @@ class UpdateRN:
                 if not integration_header:
                     rn_string += '\n### Integrations\n'
                     integration_header = True
-                rn_string += f'- __{k}__\n%%UPDATE_RN%%\n'
+                rn_string += f'#### {k}\n- %%UPDATE_RN%%\n'
             elif v == 'Playbook':
                 if not playbook_header:
                     rn_string += '\n### Playbooks\n'
                     playbook_header = True
-                rn_string += f'- __{k}__\n%%UPDATE_RN%%\n'
+                rn_string += f'#### {k}\n- %%UPDATE_RN%%\n'
             elif v == 'Script':
                 if not script_header:
                     rn_string += '\n### Scripts\n'
                     script_header = True
-                rn_string += f'- __{k}__\n%%UPDATE_RN%%\n'
+                rn_string += f'#### {k}\n- %%UPDATE_RN%%\n'
             elif v == 'IncidentFields':
                 if not inc_flds_header:
                     rn_string += '\n### IncidentFields\n'
                     inc_flds_header = True
-                rn_string += f'- __{k}__\n%%UPDATE_RN%%\n'
+                rn_string += f'#### {k}\n- %%UPDATE_RN%%\n'
             elif v == 'Classifiers':
                 if not classifier_header:
                     rn_string += '\n### Classifiers\n'
                     classifier_header = True
-                rn_string += f'- __{k}__\n%%UPDATE_RN%%\n'
+                rn_string += f'#### {k}\n- %%UPDATE_RN%%\n'
             elif v == 'Layouts':
                 if not layout_header:
                     rn_string += '\n### Layouts\n'
                     layout_header = True
-                rn_string += f'- __{k}__\n%%UPDATE_RN%%\n'
+                rn_string += f'#### {k}\n- %%UPDATE_RN%%\n'
             elif v == 'IncidentTypes':
                 if not inc_types_header:
                     rn_string += '\n### IncidentTypes\n'
                     inc_types_header = True
-                rn_string += f'- __{k}__\n%%UPDATE_RN%%\n'
+                rn_string += f'#### {k}\n- %%UPDATE_RN%%\n'
+            elif v == 'IndicatorTypes':
+                if not ind_types_header:
+                    rn_string += '\n### IndicatorTypes\n'
+                    ind_types_header = True
+                rn_string += f'#### {k}\n- %%UPDATE_RN%%\n'
         return rn_string
 
     @staticmethod
@@ -246,7 +254,7 @@ class UpdateRN:
                     continue
                 else:
                     rn_parts = new_rn.split(v + 's')
-                    new_rn_part = f'\n- __{k}__\n%%UPDATE_RN%%\n'
+                    new_rn_part = f'\n#### {k}\n- %%UPDATE_RN%%\n'
                     if len(rn_parts) > 1:
                         new_rn = rn_parts[0] + v + 's' + new_rn_part + rn_parts[1]
                     else:
@@ -254,13 +262,13 @@ class UpdateRN:
             else:
                 if v in new_rn:
                     rn_parts = new_rn.split(v + 's')
-                    new_rn_part = f'\n- __{k}__\n%%UPDATE_RN%%\n'
+                    new_rn_part = f'\n#### {k}\n- %%UPDATE_RN%%\n'
                     if len(rn_parts) > 1:
                         new_rn = rn_parts[0] + v + 's' + new_rn_part + rn_parts[1]
                     else:
                         new_rn = ''.join(rn_parts) + new_rn_part
                 else:
-                    new_rn_part = f'\n### {v}\n- __{k}__\n%%UPDATE_RN%%\n'
+                    new_rn_part = f'\n### {v}\n#### {k}\n- %%UPDATE_RN%%\n'
                     new_rn += new_rn_part
         return new_rn