+++ demisto_sdk/commands/validate/file_validator.py
@@ -20,20 +20,20 @@ import click
 import demisto_sdk.commands.common.constants as constants
 from demisto_sdk.commands.common.configuration import Configuration
 from demisto_sdk.commands.common.constants import (
-    ALL_FILES_VALIDATION_IGNORE_WHITELIST, BETA_INTEGRATION_REGEX,
-    BETA_INTEGRATION_YML_REGEX, CHECKED_TYPES_REGEXES, CODE_FILES_REGEX,
-    CONTENT_ENTITIES_DIRS, IGNORED_TYPES_REGEXES, IMAGE_REGEX,
-    INTEGRATION_REGEX, INTEGRATION_REGXES, JSON_ALL_CLASSIFIER_REGEXES,
+    ALL_FILES_VALIDATION_IGNORE_WHITELIST, CHECKED_TYPES_REGEXES,
+    CODE_FILES_REGEX, CONTENT_ENTITIES_DIRS, IGNORED_TYPES_REGEXES,
+    IMAGE_REGEX, INTEGRATION_REGXES, JSON_ALL_CLASSIFIER_REGEXES,
     JSON_ALL_CLASSIFIER_REGEXES_5_9_9, JSON_ALL_DASHBOARDS_REGEXES,
     JSON_ALL_INCIDENT_TYPES_REGEXES, JSON_ALL_INDICATOR_TYPES_REGEXES,
     JSON_ALL_LAYOUT_REGEXES, JSON_ALL_MAPPER_REGEXES,
     JSON_INDICATOR_AND_INCIDENT_FIELDS, KNOWN_FILE_STATUSES,
     OLD_YML_FORMAT_FILE, PACKAGE_SCRIPTS_REGEXES, PACKS_DIR,
-    PACKS_PACK_IGNORE_FILE_NAME, PACKS_RELEASE_NOTES_REGEX, PLAYBOOK_REGEX,
-    PLAYBOOKS_REGEXES_LIST, SCHEMA_REGEX, SCRIPT_REGEX, TEST_PLAYBOOK_REGEX,
-    YML_ALL_SCRIPTS_REGEXES, YML_BETA_INTEGRATIONS_REGEXES,
-    YML_INTEGRATION_REGEXES)
-from demisto_sdk.commands.common.errors import (ERROR_CODE,
+    PACKS_INTEGRATION_NON_SPLIT_YML_REGEX, PACKS_PACK_IGNORE_FILE_NAME,
+    PACKS_RELEASE_NOTES_REGEX, PACKS_SCRIPT_NON_SPLIT_YML_REGEX,
+    PLAYBOOK_REGEX, PLAYBOOKS_REGEXES_LIST, SCHEMA_REGEX, TEST_PLAYBOOK_REGEX,
+    YML_ALL_SCRIPTS_REGEXES, YML_INTEGRATION_REGEXES)
+from demisto_sdk.commands.common.errors import (ALLOWED_IGNORE_ERRORS,
+                                                ERROR_CODE,
                                                 FOUND_FILES_AND_ERRORS,
                                                 PRESET_ERROR_TO_CHECK,
                                                 PRESET_ERROR_TO_IGNORE, Errors)
@@ -138,7 +138,7 @@ class FilesValidator:
     def run(self):
         print_color('Starting validating files structure', LOG_COLORS.GREEN)
         if self.is_valid_structure():
-            print_color('The files are valid', LOG_COLORS.GREEN)
+            print_color('\nThe files are valid', LOG_COLORS.GREEN)
             return 0
         else:
             all_failing_files = '\n'.join(FOUND_FILES_AND_ERRORS)
@@ -365,14 +365,14 @@ class FilesValidator:
                 if not integration_validator.is_valid_file():
                     self._is_valid = False
 
-            elif checked_type(file_path, YML_BETA_INTEGRATIONS_REGEXES) or file_type == 'betaintegration':
+            elif file_type == 'betaintegration':
                 integration_validator = IntegrationValidator(structure_validator, ignored_errors=ignored_errors_list,
                                                              print_as_warnings=self.print_ignored_errors,
                                                              branch_name=self.branch_name)
                 if not integration_validator.is_valid_beta_integration():
                     self._is_valid = False
 
-            elif checked_type(file_path, [SCRIPT_REGEX]):
+            elif checked_type(file_path, [PACKS_SCRIPT_NON_SPLIT_YML_REGEX]):
                 script_validator = ScriptValidator(structure_validator, ignored_errors=ignored_errors_list,
                                                    print_as_warnings=self.print_ignored_errors,
                                                    branch_name=self.branch_name)
@@ -570,8 +570,7 @@ class FilesValidator:
                 if not script_validator.is_valid_file(validate_rn=False):
                     self._is_valid = False
 
-            elif re.match(BETA_INTEGRATION_REGEX, file_path, re.IGNORECASE) or \
-                    re.match(BETA_INTEGRATION_YML_REGEX, file_path, re.IGNORECASE) or file_type == 'betaintegration':
+            elif file_type == 'betaintegration':
                 integration_validator = IntegrationValidator(structure_validator, ignored_errors=ignored_errors_list,
                                                              print_as_warnings=self.print_ignored_errors,
                                                              branch_name=self.branch_name)
@@ -766,7 +765,7 @@ class FilesValidator:
             if not script_validator.is_valid_file(validate_rn=False):
                 self._is_valid = False
 
-        elif checked_type(file_path, YML_BETA_INTEGRATIONS_REGEXES) or file_type == 'betaintegration':
+        elif file_type == 'betaintegration':
             integration_validator = IntegrationValidator(structure_validator, ignored_errors=ignored_errors_list,
                                                          print_as_warnings=self.print_ignored_errors,
                                                          branch_name=self.branch_name)
@@ -811,12 +810,6 @@ class FilesValidator:
             if not mapper_validator.is_valid_mapper(validate_rn=False):
                 self._is_valid = False
 
-        elif checked_type(file_path, JSON_ALL_CLASSIFIER_REGEXES) or file_type == 'classifier':
-            classifier_validator = ClassifierValidator(structure_validator, ignored_errors=ignored_errors_list,
-                                                       print_as_warnings=self.print_ignored_errors)
-            if not classifier_validator.is_valid_classifier(validate_rn=False):
-                self._is_valid = False
-
         elif checked_type(file_path, JSON_ALL_CLASSIFIER_REGEXES_5_9_9) or file_type == 'classifier_5_9_9':
             classifier_validator = ClassifierValidator(structure_validator, new_classifier_version=False,
                                                        ignored_errors=ignored_errors_list,
@@ -824,8 +817,18 @@ class FilesValidator:
             if not classifier_validator.is_valid_classifier(validate_rn=False):
                 self._is_valid = False
 
+        elif checked_type(file_path, JSON_ALL_CLASSIFIER_REGEXES) or file_type == 'classifier':
+            classifier_validator = ClassifierValidator(structure_validator, ignored_errors=ignored_errors_list,
+                                                       print_as_warnings=self.print_ignored_errors)
+            if not classifier_validator.is_valid_classifier(validate_rn=False):
+                self._is_valid = False
+
+        # elif 'CHANGELOG' in file_path:
+        #     # don't check for CHANGELOG files
+        #     pass
+
         elif checked_type(file_path, CHECKED_TYPES_REGEXES):
-            print(f'Could not find validations for file {file_path}')
+            click.secho(f'Could not find validations for file {file_path}', fg='yellow')
 
         else:
             error_message, error_code = Errors.file_type_not_supported()
@@ -899,8 +902,9 @@ class FilesValidator:
         pack_files = {file for file in glob(fr'{self.file_path}/**', recursive=True) if
                       not should_file_skip_validation(file)}
         self.validate_pack_unique_files(glob(fr'{os.path.abspath(self.file_path)}'))
+        click.secho("================= Validating other pack files =================", fg="bright_cyan")
         for file in pack_files:
-            click.echo(f'Validating {file}')
+            click.echo(f'\nValidating {file}')
             self.run_all_validations_on_file(file, file_type=find_type(file))
 
     def validate_all_files_schema(self):
@@ -909,6 +913,11 @@ class FilesValidator:
         for pack_name in os.listdir(PACKS_DIR):
             pack_path = os.path.join(PACKS_DIR, pack_name)
             ignore_errors_list = self.get_error_ignore_list(pack_name)
+
+            if not os.path.isdir(pack_path):
+                # there are files that are not directories but returned from listdir like Packs/.DS_Store
+                # skip them
+                continue
 
             for dir_name in os.listdir(pack_path):
                 dir_path = os.path.join(pack_path, dir_name)
@@ -1027,12 +1036,12 @@ class FilesValidator:
     @staticmethod
     def _is_py_script_or_integration(file_path):
         file_yml = get_yaml(file_path)
-        if re.match(INTEGRATION_REGEX, file_path, re.IGNORECASE):
+        if re.match(PACKS_INTEGRATION_NON_SPLIT_YML_REGEX, file_path, re.IGNORECASE):
             if file_yml.get('script', {}).get('type', 'javascript') != 'python':
                 return False
             return True
 
-        if re.match(SCRIPT_REGEX, file_path, re.IGNORECASE):
+        if re.match(PACKS_INTEGRATION_NON_SPLIT_YML_REGEX, file_path, re.IGNORECASE):
             if file_yml.get('type', 'javascript') != 'python':
                 return False
 
@@ -1063,10 +1072,18 @@ class FilesValidator:
 
         return ignored_error_list
 
+    @staticmethod
+    def get_alowed_ignored_errors_from_list(error_list):
+        allowed_ignore_list = []
+        for error in error_list:
+            if error in ALLOWED_IGNORE_ERRORS:
+                allowed_ignore_list.append(error)
+
+        return allowed_ignore_list
+
     def add_ignored_errors_to_list(self, config, section, key, ignored_errors_list):
-        # For now one can only ignore BA101 error.
-        if key == 'ignore' and 'BA101' in str(config[section][key]).split(','):
-            ignored_errors_list.extend(['BA101'])
+        if key == 'ignore':
+            ignored_errors_list.extend(self.get_alowed_ignored_errors_from_list(str(config[section][key]).split(',')))
 
         if key in PRESET_ERROR_TO_IGNORE:
             ignored_errors_list.extend(PRESET_ERROR_TO_IGNORE.get(key))